pyproject.toml
[project]
name = "benchmarker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "jiwer>=4.0.0",
    "matplotlib>=3.10.8",
    "python-dotenv>=1.2.1",
]
